+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Update `Spinner` and `Button` components to handle responsive sizes
+-   Update `Icon` and `AvatarIcon` components to handle xl and 2xl responsive sizes
+
 ## [1.0.6] - 2023-12-13
 
 ### Added